refactor(chart): calcFallbackDistance를 Util.calcBoxDistance로 통합
